requirements.txt
beartype==0.18.5
beautifulsoup4==4.12.3
click==8.1.7
colossalai==0.3.7
core==1.0.1
deepspeed==0.15.0
diffusers==0.27.2
einops==0.8.0
ftfy==6.2.3
huggingface_hub==0.23.5
imageio==2.34.1
ipdb==0.13.13
mmcv==2.2.0
moviepy==1.0.3
numpy==2.1.1
omegaconf==2.3.0
opencv_python_headless==4.10.0.84
opensora==1.2.0
Pillow==10.4.0
pyspng==0.1.2
Requests==2.32.3
sample==1.0.0
scipy==1.14.1
torch==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
transformers==4.36.2
